requirements.txt
# Core LangChain + Community
langchain>=0.2.5,<0.4.0
langchain-community>=0.0.16,<0.4.0
langchain-core>=0.2.0,<0.4.0

# Embeddings / Vector Store
faiss-cpu>=1.7.4,<1.11.0
sentence-transformers>=2.2.2,<4.0.0
numpy>=1.24.3,<3.0.0
torch>=2.0.0,<3.0.0

# Document Loading / Processing
pypdf>=3.17.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
reportlab>=4.0.0,<5.0.0

# UI Framework (optional)
streamlit>=1.37.0,<2.0.0

# Gemini (Google Generative AI)
google-generativeai>=0.8.5